requirements.txt
asgiref==3.9.1
bleach==6.2.0
cachetools==5.5.2
certifi==2025.8.3
channels==4.3.1
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphql-core==3.2.6
graphql-relay==3.2.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
lia-web==0.2.3
opentelemetry-api==1.36.0
packaging==25.0
pillow==11.3.0
promise==2.3
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.281.0
strawberry-graphql-django==0.65.1
text-unidecode==1.3
typing_extensions==4.14.1
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
zipp==3.23.0
